requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
black==25.1.0
boto3==1.40.1
botocore==1.40.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==45.0.5
cssbeautifier==1.15.4
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-crispy-forms==2.4
django-recaptcha==4.1.0
django-storages==1.14.6
djlint==1.36.4
EditorConfig==0.17.1
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.12.0
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
mccabe==0.7.0
mutagen==1.47.0
mypy_extensions==1.1.0
openai==2.15.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-ulid==3.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
